+++ src/preconditioner.cpp
@@ -6,7 +6,7 @@ void diag_compliance(const Eigen::MatrixXd& V, const Eigen::MatrixXi& T,
     const Eigen::VectorXd& vols, double alpha,
     std::vector<Eigen::Triplet<double>>& trips) {
 
-  double offset = V.size();
+  int offset = V.size();
   for (int i = 0; i < T.rows(); ++i) {
     double He = -vols(i)/alpha;
     for (int j = 0; j < 9; ++j) {
